fix: initial version of local flow for api generator
