+++ .config/api-client-configs.json
@@ -1,7 +1,7 @@
 {
-  "bf2fc6cc711aee1a0c2a/kas-fleet-manager/kas-fleet-manager.yaml": {
-    "apiGroup": "kafka",
-    "packageName": "mgmt",
-    "apiVersion": "v1"
+  "kas-fleet-manager.yaml": {
+    "packageName": "@rhoas/kafka-manager-api",
+    "folder": "kafka-manager-api"
+
   }
 }
