+++ .github/workflows/generate_api_client.yaml
@@ -18,7 +18,8 @@ jobs:
     steps:
       - uses: actions/setup-node@v2
       - uses: actions/checkout@v2
-      - run: ./hack/scripts/generate.sh
+      - run: ./scripts/fetch_api.sh
+      - run: ./scripts/generate_js.sh
       - uses: peter-evans/create-pull-request@v3
         with:
           title: "Update API client"
