+++ .openapi-generator-ignore
@@ -1,0 +1,7 @@
+**/go.mod
+**/go.sum
+**/git_push.sh
+**/.travis.yml
+**/api
+**/.openapi-generator-ignore
+**/.openapi-generator
