+++ autosync.sh
@@ -2,6 +2,6 @@
 ## This file targets to provide list of the directories that are common across all SDKS
 
 source=../app-services-sdk-go
-folders="${source}/.chglog ${source}/.config ${source}/.github ${source}/hack"
+folders="${source}/.chglog ${source}/.config ${source}/.github ${source}/scripts"
 
 cp -r ${folders} ./
