+++ package.json
@@ -3,5 +3,8 @@
   "private": true,
   "devDependencies": {
     "lerna": "^4.0.0"
+  },
+  "scripts":{
+    "generate": "./scripts/generate_all.sh"
   }
 }
