+++ scripts/get-raw-openapi-url.js
